refactor(explorer): migrate ANHARMONICITY to PanelBuilder registry
Phase 2/8 (5th of 9 PlotTypes total).

* New explorer_internals/anharmonicity.py hosts
  AnharmonicityPanelBuilder + the absorbed
  display_anharmonicity render function.
* ANHARMONICITY has no per-plot settings UI; the builder's
  build_settings_ui returns [] (the legacy if/elif chain
  for this type fell through and implicitly returned None;
  [] is the protocol-truthful and ipyvuetify-safe equivalent).
* Registry adds PlotType.ANHARMONICITY; the dispatch branch
  is removed from Explorer.build_panel.
